Paradigm shift
So, here's the deal. We've made a whole lot of lovely progress on the helper section, but I worry that it's not informative enough, too long, and too much of an energy drain to realize it entirely. So were going in another direction. This commit is simply to keep the old code in memory, because it is absolutely dope.